pyproject.toml
[tool.poetry]
name = "quote-clock"
version = "0.1.0"
description = "A Python-based application that displays quotes on an e-paper display."
authors = ["User <user@example.com>"]
readme = "README.md"
packages = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
flask = "^2.0.0"
pillow = "^9.0.0"
pandas = "^1.5.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
